+++ package.json
@@ -8,10 +8,15 @@
   "author": "Junior Alves <junior.alves@warren.com.br>",
   "license": "MIT",
   "dependencies": {
-    "cpf-cnpj-validator": "^1.0.3",
     "express": "^4.18.2",
     "express-validator": "^6.14.2",
-    "nodemon": "^2.0.15"
+    "nodemon": "^2.0.15",
+    "typescript": "^4.9.4"
   },
-  "type": "module"
+  "type": "module",
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^5.48.1",
+    "@typescript-eslint/parser": "^5.48.1",
+    "eslint": "^8.31.0"
+  }
 }
